Move SimpleTask mock out of the crnk-test module since they're only used in crnk-openapi-gen
